+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## Changes
 
-### Version 0.16 (Next Version)
+### Version 0.16
 
 - `AudioNode` now requires `Send` and `Sync`.
 - Feedback units `Feedback64` and `Feedback32`.